+++ Editor/Domain/RenderersDomain.cs
@@ -152,7 +152,7 @@ namespace net.rs64.TexTransTool
         {
             MergeStack();
             _textureManager.DestroyDeferred();
-            _textureManager.CompressDeferred();
+            _textureManager.CompressDeferred(EnumerateRenderer(), OriginEqual);
         }
 
         public void MergeStack()